utils: remove timeout argument from nm_utils_term_child_async()
This simplifies things a little -- 5 seconds should be enough for
everybody.

It also increases the timeout it all cases. This is done so that the
timeout is also good enough for pppd, which is generally the slowest
thing to shut down.

On orderly shutdown, pp likes to tell the other side. This seems to take
at least a second even when no real network latency is at play, on busy
systems 1.5 seconds easily ends up being inadequate.

A violent pppd shutdown is generally okay apart from that it can leave
garbage (port lock) behind and the other side potentially confused for a
while.

As it happens, this interacts badly with modemu.pl which is used for
testing: the pseudo terminal in PPP line discipline mode has no idea that
the remote disconnected and while ModemManager is learning that something
wrong the hard way (AT command timing out, because the remote still expects
to talk PPP), the test times out.

Fixes-test: @gsm_sim_mtu
https://bugzilla.redhat.com/show_bug.cgi?id=2049596