libnm: fix assertion in NMClient checking for current main context
NMClient is strongly tied to the GMainContext with which it was created.
Several operations must only be called from within the context. There
was an assertion for that.

However, creating (and init_async()) should be allowed to call not
from within the GMainContext. So if the current context has no owner
(is not acquired), then it's also OK.

Fix the assertion for that.

Fixes: ce0e898fb476 ('libnm: refactor caching of D-Bus objects in NMClient')
(cherry picked from commit ae0cc9618c49bb74bbe54a073dc337e9a3b0005b)