settings-connection: expose settings plugin name on D-Bus
Add a property with the settings plugin the connection is using.
The property will be useful to identify connections that may need
migration.

The property is read-only, but a future patch will make it possible to
change a plugin the connection is using using Update2() as well as
specify a plugin on AddConnection2().