dnsmasq/shared: drop "--dhcp-lease-max=50" option
It's not clear why this option is set or why it's necessary. It was
set ever since.

Note that we automatically determine the DHCP range
nm_dnsmasq_utils_get_range(), and with /24 subnets (or larger)
the range spans up to 244 addresses.

With such a configuration it doesn't seem to make sense to limit the
number of leases.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/ ## 941